print prime factors

+++ 0x04-more_functions_nested_loops/100-prime_factor.c
@@ -8,7 +8,7 @@
 
 int main(void)
 {
-	long long prinum = 612852475143;
+	unsigned long int prinum = 612852475143;
 	int factor = 2;
 
 	while (factor * factor <= prinum)
